Update watcher-nodes.md
Watcher nodes will fail to run ABS and ABWQ synthetic transactions in TrustedServer scenario because computer account does not have access to CMS database to obtain the URL for the services, and an error will be generated in the Operations Manager event log.  Bug filed with the Product Group, and they advised this should be addressed via documentation change.  Contact me directly via email if there are questions or additional information needed.